Organize react repo from create-react-app
